requirements.txt
kafka-python>=2.0.2
pip>=24.1
setuptools>=70.1.0
wheel>=0.43.0